[#134781127] moved index.ios.js logic into a file; to share with android
